Redesigned the structure of carousel-section
Because of some difficulies with mobile view (pagination should be in the middle), have to devide slide into two parts.
Improved templates so that all slides are drawn now (which are not active - get (display: none) css class=hide).

Also the number of "dots" depend on the number of slides.

Replaced node descriptionCta with more understandable descriptionCallToAction